requirements.txt
requests==2.31.0
supabase==2.3.0
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.2.3
plotly==5.18.0
pytest==7.4.3
pytest-cov==4.1.0

